pyproject.toml
[tool.poetry]
name = "biblioteca"
version = "0.1.0"
description = ""
authors = ["Keane Moraes <lordvader3002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
segment-anything = "^1.0"
diffusers = "^0.35.1"
transformers = "^4.56.1"
accelerate = "^1.10.1"
scipy = "^1.16.1"
safetensors = "^0.6.2"
opencv-python = "4.10.0.82"
matplotlib = "^3.10.6"
pillow = "^11.3.0"
torchvision = "0.17.1"
tqdm = "^4.67.1"
jupyter = "^1.1.1"
groundingdino-py = "^0.4.0"
hf-transfer = "^0.1.9"
numpy = "1.26.0"
torch = "2.2.1"
supervision = "0.6.0"
easyocr = "^1.7.2"
openai = "^1.109.1"
pydantic = "^2.11.9"
ultralytics = "^8.3.203"
wandb = "^0.22.0"
weave = "^0.52.8"
requests = "^2.32.5"
ipykernel = "^6.30.1"
fastapi = "^0.118.0"
uvicorn = "^0.37.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
